feat(middleware): FT(main) -  Active skip token

+++ order-management-asp-net/OrderManagement.Application/Interfaces/Custom/IJWTService.cs
@@ -8,7 +8,7 @@ public interface IJWTService
 
     public string GetUserCode(string token);
 
-    public long GetUserId(string token);
+    public int GetUserId(string token);
 
     public bool ValidateToken(string token);
 }
